requirements.txt
matplotlib 
numpy
scipy 
tqdm 
pandas 
argparse 
PyYAML 
spatialmath-python
[-e] git+https://gitlab.aau.at/aau-cns/py3_pkgs/numpy_utils.git#egg=numpy_utils
[-e] git+https://gitlab.aau.at/aau-cns/py3_pkgs/trajectory.git#egg=trajectory
[-e] git+https://gitlab.aau.at/aau-cns/py3_pkgs/csv2dataframe.git#egg=csv2dataframe
[-e] git+https://gitlab.aau.at/aau-cns/py3_pkgs/ros_csv_formats.git#egg=ros_csv_formats




